==> R/10_plot_min_coral_macro.R <==
# load packages and the data we need ===========================================
library(scales)
library(deSolve)
library(tidyverse)
library(readxl)
library(here)
library(qs)
library(viridis)

source(here::here("./R/00_global_funs.R"))

`%notin%` <- Negate(`%in%`)

matching_df <- readr::read_csv(
  here::here("./data/plotting-data/min-coral-ids-macro.csv")
) %>%
  dplyr::mutate(macro = factor(macro, levels = c("low", "med", "high")))

# set up the data for the different levels =====================================
matching_df <- matching_df %>% dplyr::mutate(
  min_coral = ifelse(min_coral != -999999, min_coral, NA)
)
matching_df$grazing_level <- NA
matching_df[which(matching_df$g <= 0.2), "grazing_level"] <- "0-0.2"
matching_df[which(matching_df$g > 0.2 &
  matching_df$g <= 0.4), "grazing_level"] <- "0.2-0.4"
matching_df[which(matching_df$g > 0.4), "grazing_level"] <- ">0.4"
matching_df$grazing_level <- factor(matching_df$grazing_level,
  levels = c("0-0.2", "0.2-0.4", ">0.4")
)

matching_df$overgrow_level <- NA
matching_df[which(matching_df$a <= 0.33), "overgrow_level"] <- "low"
matching_df[which(matching_df$a > 0.33 &
  matching_df$a <= 0.66), "overgrow_level"] <- "med"
matching_df[which(matching_df$a > 0.66), "overgrow_level"] <- "high"
matching_df$overgrow_level <- factor(matching_df$overgrow_level,
  levels = c("low", "med", "high")
)

matching_df$recruit_level <- NA
matching_df[which(matching_df$z <= 0.33), "recruit_level"] <- "low"
matching_df[which(matching_df$z > 0.33 &
  matching_df$z <= 0.66), "recruit_level"] <- "med"
matching_df[which(matching_df$z > 0.66), "recruit_level"] <- "high"
matching_df$recruit_level <- factor(matching_df$recruit_level,
  levels = c("low", "med", "high")
)

# make the plots ===============================================================
generate_colors <- function(n, high, mid, low) {
  # Create a color ramp palette function using high, mid, and low colors
  color_palette <- colorRampPalette(c(low, mid, high))

  # Generate a sequence of n colors
  colors <- color_palette(n)

  return(colors)
}

# Example usage
n <- 182
low <- "#16665f" # Red for the highest value
mid <- "#e7f57d" # White for the midpoint
high <- "#9b4804" # Blue for the lowest value

cols <- generate_colors(n, high, mid, low)

div_by_macro_recruit <- ggplot(data = matching_df) +
  geom_point(aes(x = a, y = g, colour = min_coral, fill = "NA"),
    alpha = 0.3, size = 2
  ) +
  facet_grid(macro ~ recruit_level) +
  theme_base() +
  scale_color_gradientn("Minimum Coral",
    breaks = c(0.01, 0.25, 0.5, 0.75, 0.96),
    colors = cols,
    na.value = "grey80"
  ) +
  labs(x = "Macroalgal-Coral Comp.", y = "Grazing") +
  scale_x_continuous(
    breaks = c(0.01, 0.25, 0.5, 0.75, 0.99),
    labels = c(0.01, 0.25, 0.5, 0.75, 1.0),
    sec.axis = sec_axis(~.,
      name = "Recruitment",
      breaks = NULL, labels = NULL
    )
  ) +
  scale_y_continuous(
    breaks = c(0.01, 0.25, 0.5, 0.75, 0.99),
    labels = c(0.01, 0.25, 0.5, 0.75, 1.0),
    sec.axis = sec_axis(~.,
      name = "Macroalgae Level",
      breaks = NULL, labels = NULL
    )
  ) +
  scale_fill_manual(values = "grey80", labels = "") +
  guides(
    colour = guide_colorbar(
      title = "Minimum Coral",
      order = 1
    ),
    fill = guide_legend(
      title = "No Possible Combination",
      override.aes = list(
        fill = "grey80",
        size = 4,
        order = 2
      )
    )
  )
ggsave(
  here::here("./graphs/conclusions-plots/min-coral-by-macro-div-by-recruit.png"),
  div_by_macro_recruit,
  height = 6, width = 10
)
div_by_macro_comp <- ggplot(data = matching_df) +
  geom_point(aes(x = z, y = g, colour = min_coral, fill = "NA"),
    alpha = 0.3, size = 2
  ) +
  facet_grid(macro ~ overgrow_level) +
  theme_base() +
  scale_color_gradientn("Minimum Coral",
    breaks = c(0.01, 0.25, 0.5, 0.75, 0.96),
    colors = cols,
    na.value = "grey80"
  ) +
  labs(x = "Recruitment", y = "Grazing") +
  scale_x_continuous(
    breaks = c(0, 0.25, 0.5, 0.75, 0.99),
    labels = c(0, 0.25, 0.5, 0.75, 1.0),
    sec.axis = sec_axis(~.,
      name = "Macroalgal-Coral Comp.",
      breaks = NULL, labels = NULL
    )
  ) +
  scale_y_continuous(
    breaks = c(0.01, 0.25, 0.5, 0.75, 0.99),
    labels = c(0.01, 0.25, 0.5, 0.75, 1.0),
    sec.axis = sec_axis(~.,
      name = "Macroalgae Level",
      breaks = NULL, labels = NULL
    )
  ) +
  scale_fill_manual(values = "grey80", labels = "") +
  guides(
    colour = guide_colorbar(
      title = "Minimum Coral",
      order = 1
    ),
    fill = guide_legend(
      title = "No Possible Combination",
      override.aes = list(
        fill = "grey80",
        size = 4,
        order = 2
      )
    )
  )
ggsave(
  here::here("./graphs/conclusions-plots/min-coral-by-macro-div-by-comp.png"),
  div_by_macro_comp,
  height = 6, width = 10
)

div_by_macro_grazing <- ggplot(data = matching_df) +
  geom_point(aes(x = a, y = z, colour = min_coral, fill = "NA"),
    alpha = 0.3, size = 2
  ) +
  facet_grid(macro ~ grazing_level) +
  theme_base() +
  scale_color_gradientn("Minimum Coral",
    breaks = c(0.01, 0.25, 0.5, 0.75, 0.96),
    colors = cols,
    na.value = "grey80"
  ) +
  labs(x = "Macroalgal-Coral Comp.", y = "Recruitment") +
  scale_x_continuous(
    breaks = c(0, 0.25, 0.5, 0.75, 0.99),
    labels = c(0, 0.25, 0.5, 0.75, 1.0),
    sec.axis = sec_axis(~.,
      name = "Grazing",
      breaks = NULL, labels = NULL
    )
  ) +
  scale_y_continuous(
    breaks = c(0.01, 0.25, 0.5, 0.75, 0.99),
    labels = c(0.01, 0.25, 0.5, 0.75, 1.0),
    sec.axis = sec_axis(~.,
      name = "Macroalgae Level",
      breaks = NULL, labels = NULL
    )
  ) +
  scale_fill_manual(values = "grey80", labels = "") +
  guides(
    colour = guide_colorbar(
      title = "Minimum Coral",
      order = 1
    ),
    fill = guide_legend(
      title = "No Possible Combination",
      override.aes = list(
        fill = "grey80",
        size = 4,
        order = 2
      )
    )
  )

ggsave(
  here::here("./graphs/conclusions-plots/min-coral-by-macro-div-by-grazing.png"),
  div_by_macro_grazing,
  height = 6, width = 10
)
