==> R/15_all_plotting.R <==
#' I just need a single file to run if I make plotting updates because that
#' will make my life easier lol

source(here::here("./R/07_plot_min_coral.R"))
source(here::here("./R/08_plot_goodID_props.R"))
source(here::here("./R/10_plot_min_coral_macro.R"))
source(here::here("./R/13_plot_all_min_coral.R"))
source(here::here("./R/14_01_trajectories_plotting_nhrcp.R"))
source(here::here("./R/14_02_trajectories_plotting_tnc.R"))
